Alterados tests de CRUD de Product
