+++ liberaction/core/templates/core/product.html
@@ -1,6 +1,6 @@
-<h1>{{ product.name }}</h1>
-<h5>{{ product.owner }}</h5>
-<h5>{{ product.description }}</h5>
+<h1>{{ product.get_name }}</h1>
+<h5>{{ product.get_owner }}</h5>
+<h5>{{ product.get_description }}</h5>
 
 {% for picture in product.get_pictures %}
     <img src="{{ picture.img.url }}" alt="Foto #{{ picture.index }}">
